requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
astrapy==2.0.1
asttokens==3.0.0
asyncer==0.0.7
asyncpg==0.30.0
attrs==25.3.0
autopep8==2.3.2
backoff==2.2.1
beautifulsoup4==4.13.3
bidict==0.23.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chainlit==2.5.5
chardet==5.2.0
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cohere==5.11.3
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
emoji==2.14.1
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
infisical-python==2.3.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.7
langchain-astradb==0.6.0
langchain-cohere==0.3.1
langchain-community==0.3.5
langchain-core==0.3.15
langchain-experimental==0.3.2
langchain-google-genai==2.1.2
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.139
Lazify==0.4.0
literalai==0.1.201
lxml==5.3.2
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.6.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
olefile==0.47
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.39.1
opentelemetry-instrumentation-anthropic==0.39.1
opentelemetry-instrumentation-bedrock==0.39.1
opentelemetry-instrumentation-chromadb==0.39.1
opentelemetry-instrumentation-cohere==0.39.1
opentelemetry-instrumentation-crewai==0.39.1
opentelemetry-instrumentation-google-generativeai==0.39.1
opentelemetry-instrumentation-groq==0.39.1
opentelemetry-instrumentation-haystack==0.39.1
opentelemetry-instrumentation-lancedb==0.39.1
opentelemetry-instrumentation-langchain==0.39.1
opentelemetry-instrumentation-llamaindex==0.39.1
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.39.1
opentelemetry-instrumentation-milvus==0.39.1
opentelemetry-instrumentation-mistralai==0.39.1
opentelemetry-instrumentation-ollama==0.39.1
opentelemetry-instrumentation-openai==0.39.1
opentelemetry-instrumentation-pinecone==0.39.1
opentelemetry-instrumentation-qdrant==0.39.1
opentelemetry-instrumentation-replicate==0.39.1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.39.1
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.39.1
opentelemetry-instrumentation-transformers==0.39.1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.39.1
opentelemetry-instrumentation-watsonx==0.39.1
opentelemetry-instrumentation-weaviate==0.39.1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.52b1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
posthog==3.24.1
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.12.0
PyMuPDF==1.24.13
pypdf==5.4.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-engineio==4.12.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.18
python-oxmsg==0.0.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
syncer==2.0.3
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.2.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.6.0+cpu
tornado==6.4.2
tqdm==4.66.6
traceloop-sdk==0.39.1
traitlets==5.14.3
transformers==4.46.3
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
unstructured==0.16.4
unstructured-client==0.26.2
update-checker==0.18.0
uptrace==1.31.0
urllib3==2.3.0
uuid6==2024.7.10
uvicorn==0.34.1
watchfiles==0.20.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
